refactor: 리펙토링 사항 반영
