+++ src/main/java/com/spaceclub/global/config/RestTemplateConfig.java
@@ -1,0 +1,15 @@
+package com.spaceclub.global.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestTemplateConfig {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+
+}
