+++ src/main/java/com/spaceclub/global/oauth/config/KakaoOauthInfoSender.java
@@ -1,12 +1,11 @@
-package com.spaceclub.global.oauth.service;
+package com.spaceclub.global.oauth.config;
 
-import com.spaceclub.global.oauth.config.KakaoOauthProperties;
-import com.spaceclub.global.oauth.service.vo.KakaoTokenInfo;
-import com.spaceclub.global.oauth.service.vo.KakaoUserInfo;
+import com.spaceclub.global.oauth.config.vo.KakaoTokenInfo;
+import com.spaceclub.global.oauth.config.vo.KakaoUserInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
@@ -14,23 +13,24 @@ import org.springframework.web.client.RestTemplate;
 import static org.springframework.http.HttpMethod.POST;
 import static org.springframework.http.MediaType.APPLICATION_FORM_URLENCODED;
 
-@Service
+@Component
 @RequiredArgsConstructor
-public class OAuthService {
+public class KakaoOauthInfoSender {
 
     private static final String REQUEST_TOKEN_URL = "https://kauth.kakao.com/oauth/token";
     private static final String REQUEST_INFO_URL = "https://kapi.kakao.com/v2/user/me";
     private static final String GRANT_TYPE = "authorization_code";
 
+    private final RestTemplate restTemplate;
     private final KakaoOauthProperties kakaoProperties;
 
     public KakaoUserInfo getUserInfo(String accessToken) {
-        RestTemplate restTemplate = new RestTemplate();
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.setContentType(APPLICATION_FORM_URLENCODED);
         httpHeaders.setBearerAuth(accessToken);
 
         HttpEntity<Object> requestEntity = new HttpEntity<>(httpHeaders);
+
         return restTemplate.exchange(
                 REQUEST_INFO_URL,
                 POST,
@@ -40,7 +40,6 @@ public class OAuthService {
     }
 
     public KakaoTokenInfo getAccessTokenInfo(String code) {
-        RestTemplate restTemplate = new RestTemplate();
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.setContentType(APPLICATION_FORM_URLENCODED);
 
@@ -52,6 +51,7 @@ public class OAuthService {
         body.add("code", code);
 
         HttpEntity<Object> requestEntity = new HttpEntity<>(body, httpHeaders);
+
         return restTemplate.exchange(
                 REQUEST_TOKEN_URL,
                 POST,
