+++ src/main/java/com/spaceclub/global/oauth/config/vo/KakaoUserInfo.java
@@ -1,4 +1,4 @@
-package com.spaceclub.global.oauth.service.vo;
+package com.spaceclub.global.oauth.config.vo;
 
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.spaceclub.user.domain.Status;
