+++ src/main/java/com/spaceclub/user/controller/UserController.java
@@ -2,9 +2,6 @@ package com.spaceclub.user.controller;
 
 import com.spaceclub.event.domain.Event;
 import com.spaceclub.global.dto.PageResponse;
-import com.spaceclub.global.oauth.service.OAuthService;
-import com.spaceclub.global.oauth.service.vo.KakaoTokenInfo;
-import com.spaceclub.global.oauth.service.vo.KakaoUserInfo;
 import com.spaceclub.user.controller.dto.UserEventGetResponse;
 import com.spaceclub.user.controller.dto.UserLoginResponse;
 import com.spaceclub.user.domain.User;
@@ -29,7 +26,6 @@ import static com.spaceclub.user.controller.dto.UserEventGetResponse.from;
 public class UserController {
 
     private final UserService userService;
-    private final OAuthService oAuthService;
 
     @GetMapping("/{userId}/events")
     public PageResponse<UserEventGetResponse, Event> getAllEvents(@PathVariable Long userId, Pageable pageable) {
@@ -44,18 +40,8 @@ public class UserController {
 
     @PostMapping("/oauths")
     public UserLoginResponse getKaKaoCode(@RequestParam String code) {
-        User user = createKakaoUser(code);
-        userService.save(user);
-        boolean newMember = userService.isNewMember(user);
-
-        return new UserLoginResponse("accessToken", newMember);
-    }
-
-    private User createKakaoUser(String code) {
-        KakaoTokenInfo accessTokenInfo = oAuthService.getAccessTokenInfo(code);
-        String accessToken = accessTokenInfo.accessToken();
-        KakaoUserInfo userInfo = oAuthService.getUserInfo(accessToken);
-        return userInfo.toUser();
+        User kakaoUser = userService.createKakaoUser(code);
+        return UserLoginResponse.from(kakaoUser);
     }
 
 }
