+++ src/main/java/com/spaceclub/user/controller/dto/UserLoginResponse.java
@@ -1,5 +1,11 @@
 package com.spaceclub.user.controller.dto;
 
+import com.spaceclub.user.domain.User;
+
 public record UserLoginResponse(String accessToken, boolean isNewMember) {
+
+    public static UserLoginResponse from(User user) {
+        return new UserLoginResponse("accessToken",user.isNewMember());
+    }
 
 }
